+++ bin/create-tool.sh
@@ -2,7 +2,7 @@
 # build-containers
 #
 # Create a blank tool.
-VERSION=1.4.1
+VERSION=1.4.2
 
 if [[ $# == 0 ]]; then
     echo ""