+++ bin/helpers/bactopia-build.py
@@ -21,7 +21,7 @@ optional arguments:
 import logging
 import os
 
-VERSION = "1.4.1"
+VERSION = "1.4.2"
 PROGRAM = "bactopia build"
 STDOUT = 11
 STDERR = 12