+++ tools/phyloflash/bin/jsonify-phyloflash.py
@@ -15,7 +15,7 @@ example usage:
   jsonify-phyloflash ./
 """
 PROGRAM = "jsonify-phyloflash"
-VERSION = "1.4.1"
+VERSION = "1.4.2"
 
 
 def read_phyloflash(phyloflash):